Updated mersenne check, to use Lucas-Lehmer test from https://en.wikipedia.org/wiki/Lucas%E2%80%93Lehmer_primality_test

+++ Charitas.Processes/Interfaces/IMersennePrimes.cs
@@ -4,8 +4,8 @@ namespace Charitas.Processes.Interfaces
 {
     public interface IMersennePrimes
     {
-        Task FindMersennePrimes(long start);
+        Task FindMersennePrimes(int start);
 
-        bool IsMersennePrime(long candidate);
+        bool IsMersennePrime(int candidate);
     }
 }
